refactor: 모달 동작 관리 로직 훅으로 분리
